Move BackupsList and listItems into seperate component
